code cleanup

+++ src/main/java/calc/tokenizer/token/OperatorADD.java
@@ -16,11 +16,11 @@ public class OperatorADD implements OperatorType {
     }
 
     public NumberType operate(IntegerType op1, IntegerType op2) {
-        return new IntegerType(op1.value.intValue() + op2.value.intValue());
+        return new IntegerType(op1.value + op2.value);
     }
 
     public NumberType operate(DoubleType op1, IntegerType op2) {
-        return new DoubleType(op1.value.doubleValue() + op2.value.intValue());
+        return new DoubleType(op1.value + op2.value);
     }
 
     public NumberType operate(IntegerType op1, DoubleType op2) {
@@ -28,7 +28,7 @@ public class OperatorADD implements OperatorType {
     }
 
     public NumberType operate(DoubleType op1, DoubleType op2) {
-        return new DoubleType(op1.value.doubleValue() + op2.value.doubleValue());
+        return new DoubleType(op1.value + op2.value);
     }
 
     public String toString() {
