+++ src/main/java/calc/tokenizer/token/OperatorDIV.java
@@ -16,19 +16,19 @@ public class OperatorDIV implements OperatorType {
     }
 
     public DoubleType operate(IntegerType op1, IntegerType op2) {
-        return new DoubleType((double) op1.value.intValue() / (double) op2.value.intValue());
+        return new DoubleType((double) op1.value / (double) op2.value);
     }
 
     public DoubleType operate(DoubleType op1, DoubleType op2) {
-        return new DoubleType(op1.value.doubleValue() / op2.value.doubleValue());
+        return new DoubleType(op1.value / op2.value);
     }
 
     public DoubleType operate(IntegerType op1, DoubleType op2) {
-        return new DoubleType(op1.value.intValue() / op2.value.doubleValue());
+        return new DoubleType(op1.value / op2.value);
     }
 
     public DoubleType operate(DoubleType op1, IntegerType op2) {
-        return new DoubleType(op1.value.doubleValue() / op2.value.intValue());
+        return new DoubleType(op1.value / op2.value);
     }
 
     public String toString() {
