+++ src/main/java/calc/tokenizer/token/OperatorPOW.java
@@ -1,7 +1,5 @@
 package calc.tokenizer.token;
 
-import calc.tokenizer.token.IntegerType;
-import calc.tokenizer.token.DoubleType;
 import calc.tokenizer.token.type.NumberType;
 import calc.tokenizer.token.type.OperatorType;
 
@@ -20,15 +18,15 @@ public class OperatorPOW implements OperatorType {
     }
 
     public DoubleType operate(DoubleType op1, DoubleType op2) {
-        return new DoubleType(Math.pow(op1.value.doubleValue(), op2.value.doubleValue()));
+        return new DoubleType(Math.pow(op1.value, op2.value));
     }
 
     public DoubleType operate(IntegerType op1, DoubleType op2) {
-        return new DoubleType(Math.pow(op1.value.intValue(), op2.value.doubleValue()));
+        return new DoubleType(Math.pow(op1.value, op2.value));
     }
 
     public DoubleType operate(DoubleType op1, IntegerType op2) {
-        return new DoubleType(Math.pow(op1.value.doubleValue(), op2.value.doubleValue()));
+        return new DoubleType(Math.pow(op1.value, op2.value.doubleValue()));
     }
 
     public String toString() {
