+++ src/main/java/calc/tokenizer/token/OperatorSUB.java
@@ -1,7 +1,5 @@
 package calc.tokenizer.token;
 
-import calc.tokenizer.token.IntegerType;
-import calc.tokenizer.token.DoubleType;
 import calc.tokenizer.token.type.NumberType;
 import calc.tokenizer.token.type.OperatorType;
 
@@ -16,19 +14,19 @@ public class OperatorSUB implements OperatorType {
     }
 
     public NumberType operate(IntegerType op1, IntegerType op2) {
-        return new IntegerType(op1.value.intValue() - op2.value.intValue());
+        return new IntegerType(op1.value - op2.value);
     }
 
     public NumberType operate(DoubleType op1, IntegerType op2) {
-        return new DoubleType(op1.value.doubleValue() - op2.value.intValue());
+        return new DoubleType(op1.value - op2.value);
     }
 
     public NumberType operate(IntegerType op1, DoubleType op2) {
-        return new DoubleType(op1.value.intValue() - op2.value.doubleValue());
+        return new DoubleType(op1.value - op2.value);
     }
 
     public NumberType operate(DoubleType op1, DoubleType op2) {
-        return new DoubleType(op1.value.doubleValue() - op2.value.doubleValue());
+        return new DoubleType(op1.value - op2.value);
     }
 
     public String toString() {
